# src/main/java/fitnesstracker/controllers/HealthStatisticController.java
package fitnesstracker.controllers;

import fitnesstracker.entities.health.HealthStatistic;
import fitnesstracker.services.HealthStatisticService;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.responses.ApiResponses;
import io.swagger.v3.oas.annotations.tags.Tag;
import org.jetbrains.annotations.NotNull;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.http.HttpStatus;
import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.*;

import java.util.List;

@RestController
@RequestMapping("/stats")
@Tag(name = "Health Statistics", description = "Health Statistics APIs")
public class HealthStatisticController {

    private final HealthStatisticService healthStatisticService ;

    @Autowired
    public HealthStatisticController(HealthStatisticService healthStatisticService) {
        this.healthStatisticService = healthStatisticService;
    }

    @GetMapping
    @Operation(summary = "Get all Health Statistics", description = "Returns all Health Statistics as per the id",
    tags = {"stats", "get"})
    @ApiResponses(value = {
            @ApiResponse(responseCode = "200", content = { @Content(schema = @Schema(implementation = HealthStatistic.class), mediaType = "application/json") }, description = "Successfully retrieved"),
            @ApiResponse(responseCode = "404", description = "Not found - The Health Statistics were not found", content = { @Content(schema = @Schema()) })
    })
    public List<HealthStatistic> getAllHealthStatistics(
    ) {
        return healthStatisticService.getAllHealthStatistics();
    }


    @GetMapping("/{id}")
    public ResponseEntity<HealthStatistic> getHealthStatisticById(@PathVariable Long id) {
        HealthStatistic healthStatistic = healthStatisticService.getHealthStatisticById(id);

        if (healthStatistic != null) {
            return ResponseEntity.ok(healthStatistic);
        } else {
            return ResponseEntity.notFound().build();
        }
    }

    @PostMapping
    public ResponseEntity<HealthStatistic> createHealthStatistic(@RequestBody @NotNull HealthStatistic healthStatistic) {
        @SuppressWarnings("unused")
        Long ignoredPersonId = healthStatistic.getPersonId();
        HealthStatistic createdHealthStatistic = healthStatisticService.createHealthStatistic(healthStatistic);
        return ResponseEntity.status(HttpStatus.CREATED).body(createdHealthStatistic);
    }

    @DeleteMapping("/{id}")
    public ResponseEntity<Void> deleteHealthStatistic(@PathVariable Long id) {
        healthStatisticService.deleteHealthStatistic(id);
        return ResponseEntity.noContent().build();
    }
}