# src/main/resources/openapi.json
{"openapi":"3.0.1","info":{"title":"OpenAPI definition","version":"v0"},"servers":[{"url":"http://localhost:8081","description":"Generated server url"}],"tags":[{"name":"Meals","description":"Meal APIs"},{"name":"Health Statistics","description":"Health Statistics APIs"}],"paths":{"/stats":{"get":{"tags":["Health Statistics","stats","get"],"summary":"Get all Health Statistics","description":"Returns all Health Statistics as per the id","operationId":"getAllHealthStatistics","responses":{"404":{"description":"Not found - The Health Statistics were not found"},"200":{"description":"Successfully retrieved","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HealthStatistic"}}}}}},"post":{"tags":["Health Statistics"],"operationId":"createHealthStatistic","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/HealthStatistic"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/HealthStatistic"}}}}}}},"/meals":{"get":{"tags":["get","Meals","meals"],"summary":"Get all Meals","description":"Returns a list of all Meals","operationId":"getAllMeals","responses":{"500":{"description":"Internal Server Error"},"200":{"description":"Successfully retrieved","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Meal"}}}}}},"post":{"tags":["post","Meals","meals"],"summary":"Add a new Meal","description":"Adds a new Meal","operationId":"addMeal","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Meal"}}},"required":true},"responses":{"201":{"description":"Successfully created","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Meal"}}}},"400":{"description":"Bad Request"}}}},"/exercises":{"get":{"tags":["exercise-history-controller"],"operationId":"getAllExercises","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}},"post":{"tags":["exercise-history-controller"],"operationId":"addNewExercise","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Exercise"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/stats/{id}":{"get":{"tags":["Health Statistics"],"operationId":"getHealthStatisticById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/HealthStatistic"}}}}}},"delete":{"tags":["Health Statistics"],"operationId":"deleteHealthStatistic","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK"}}}},"/meals/{mealId}":{"get":{"tags":["get","Meals","meals"],"summary":"Get a Meal by ID","description":"Returns a Meal based on its ID","operationId":"getMealById","parameters":[{"name":"mealId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"404":{"description":"Not Found"},"200":{"description":"Successfully retrieved","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Meal"}}}}}}},"/exercises/{exerciseId}":{"get":{"tags":["exercise-history-controller"],"operationId":"getExerciseById","parameters":[{"name":"exerciseId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Exercise"}}}}}}},"/exercises/person/{personId}":{"get":{"tags":["exercise-history-controller"],"operationId":"findByPersonId","parameters":[{"name":"personId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Exercise"}}}}}}}},"/exercises/pb/test":{"get":{"tags":["personal-best-controller"],"operationId":"getBestDeadlifts","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Exercise"}}}}}}},"/exercises/name/{exerciseName}":{"get":{"tags":["exercise-history-controller"],"operationId":"getExerciseByName","parameters":[{"name":"exerciseName","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Exercise"}}}}}}}},"/exercises/{id}":{"delete":{"tags":["exercise-history-controller"],"operationId":"deleteExercise","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK"}}}}},"components":{"schemas":{"HealthStatistic":{"type":"object","properties":{"id":{"type":"integer","description":"Health Statistics ID","format":"int64","readOnly":true,"example":1},"date":{"type":"string","description":"Date of the Health Statistics","format":"date","example":"2020-01-01"},"sleep":{"type":"number","description":"Sleep of the Health Statistics","format":"double","example":10},"weight":{"type":"number","description":"Weight of the Health Statistics","format":"double","example":10},"bloodPressure":{"type":"number","format":"double"},"hydration":{"type":"number","format":"double"},"calorieIn":{"type":"integer","format":"int32"},"calorieOut":{"type":"number","format":"double"},"heartRate":{"type":"number","format":"double"},"stress":{"type":"number","format":"double"},"personId":{"type":"integer","format":"int64"}},"description":"Health Statistics"},"Ingredient":{"type":"object","properties":{"id":{"type":"integer","description":"Ingredient ID","format":"int64","readOnly":true,"example":1},"ingredientName":{"type":"string","description":"Name of the Ingredient","example":"Banana"}},"description":"Ingredient Information"},"Meal":{"type":"object","properties":{"id":{"type":"integer","description":"Meal ID","format":"int64","readOnly":true,"example":1},"date":{"type":"string","description":"Date of the Meal","format":"date","example":"2020-01-01"},"mealName":{"type":"string","description":"Name of the Meal","example":"Beans on Toast"},"mealType":{"type":"string","description":"Type of the Meal","example":"Breakfast"},"calories":{"type":"integer","description":"Calories of the Meal","format":"int32","example":1000},"cookingInstructions":{"type":"string","description":"Cooking instructions for the Meal","example":"Bake 100g of beans, put toast in toaster"},"ingredientList":{"type":"array","items":{"$ref":"#/components/schemas/Ingredient"}},"personId":{"type":"integer","format":"int64"}},"description":"Meal Information"},"Exercise":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"exerciseName":{"type":"string"},"equipmentRequired":{"type":"string"},"startTime":{"type":"string","format":"date-time"},"endTime":{"type":"string","format":"date-time"},"caloriesBurned":{"type":"integer","format":"int32"},"description":{"type":"string"},"duration":{"type":"object","properties":{"seconds":{"type":"integer","format":"int64"},"zero":{"type":"boolean"},"nano":{"type":"integer","format":"int32"},"negative":{"type":"boolean"},"units":{"type":"array","items":{"type":"object","properties":{"durationEstimated":{"type":"boolean"},"timeBased":{"type":"boolean"},"dateBased":{"type":"boolean"}}}}}},"personId":{"type":"integer","format":"int64"}}}}}}